requirements.txt
rumps>=0.4.0
sounddevice>=0.4.6
numpy>=1.24.0
scipy>=1.10.0
faster-whisper>=1.0.0
google-genai>=1.0.0
pyperclip>=1.8.0
pyobjc-framework-Quartz>=9.0
